refactor(lambda): remove duplciate utils
