+++ backend/lambda/utils/schedule.utils.ts
