--- backend/lambda/utils/validators.ts
@@ -1,3 +1,0 @@
-export function isIsoDateFormat(date: string) {
-  return date.length === 10 && !Number.isNaN(new Date(date).getTime());
-}
